requirements.txt
python-dotenv==1.0.1
streamlit==1.41.1
huggingface-hub==0.29.1
openai==1.65.2
langchain
PyPDF2==3.0.1
faiss-cpu==1.10.0
langchain-community
PyMySQL==1.1.1
tiktoken==0.9.0